feat: Moved autocomplete to dedicated class file
